% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_discretes.R
\name{as_discretes}
\alias{as_discretes}
\title{Create a discrete series from a numeric vector.}
\usage{
as_discretes(x)
}
\arguments{
\item{x}{A numeric vector with no missing values. May contain infinity.}
}
\value{
A "discretes" object whose members are the unique values of \code{x}.
}
\description{
Create a discrete series from a numeric vector.
}
\examples{
as_discretes(0:10)
}
