% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{as_integerish}
\alias{as_integerish}
\title{"Safely" convert to integer-like}
\usage{
as_integerish(x)
}
\arguments{
\item{x}{Numeric vector. More specifically, an atomic vector that is
coercable to numeric via \code{as.numeric()} without becoming \code{NA}.}
}
\value{
A vector comprising of whole numbers, with type "integer" if
possible, but at least type "double".
}
\description{
R doesn't allow numbers to hold an integer type if they are too big.
\code{as.integer()} will convert such numbers to \code{NA}, so that its output
is always of type "integer". \code{as_integerish()}, on the other hand,
would rather keep the original number and change the type to "double"
than to convert large numbers to \code{NA}.
}
\details{
When \code{as.integer()} coneverts a number to \code{NA}, \code{as_integerish()} will
instead drop the fractional part of the number, keeping the sign.

Values in \code{x} that are \code{NA} are preserved as \code{NA}.

\code{as.integer()} is quite forgiving with its inputs.
As of 'base' version 4.4.3, \code{as.integer("5.5")} returns \code{5L}, for example.
\code{as_integerish()} adopts the same behaviour by allowing inputs that don't
get converted to \code{NA} by \code{as.numeric()}.
}
\examples{
suppressWarnings(as.integer(1e100))
discretes:::as_integerish(1e100)
discretes:::as_integerish(1e100 + 0.8) == 1e100
discretes:::as_integerish(-1e100 - 0.8) == -1e100

x <- discretes:::as_integerish(c(NA, 1, -5.5))
is.integer(x)

y <- discretes:::as_integerish(c(NA, 1, -5.5, 1e100))
is.integer(y)
}
