% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsct_empty.R
\name{dsct_empty}
\alias{dsct_empty}
\title{Create an empty discrete set}
\usage{
dsct_empty(mode = c("double", "integer"))
}
\arguments{
\item{mode}{Character vector of length 1 indicating the numeric type of the
discrete set; either "double" (the default) or "integer".}
}
\value{
An empty discrete set of the specified mode.
}
\description{
Create a discrete set with no members.
}
\examples{
dsct_empty()
}
