% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsct_invert.R
\name{dsct_invert}
\alias{dsct_invert}
\title{Invert a Discrete Value Series}
\usage{
dsct_invert(x)
}
\arguments{
\item{x}{Discrete value series
(\code{numeric} vector or object of class \code{"discretes"}).}
}
\value{
A new discrete value series containing the reciprocals of all values
in \code{x}.
}
\description{
Create a discrete value series by taking the reciprocal of values in
another discrete value series.
}
\details{
By default, this transformation creates a new \code{"dsct_inverse"} subclass by
wrapping the input object. Simplifications occur when behaviour is known,
for the following series:
\itemize{
\item \strong{Numeric vectors}: The reciprocal of each value is computed directly,
returning another numeric vector.
\item \strong{Inverse series}: The inverse of an inverse series returns the original
series.
}
}
\examples{
dsct_invert()

}
\seealso{
Other transformations: 
\code{\link{dsct_drop}()},
\code{\link{dsct_negate}()},
\code{\link{dsct_transform}()},
\code{\link{dsct_union}()}
}
\concept{transformations}
