% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsct_linear.R
\name{dsct_linear}
\alias{dsct_linear}
\title{Linearly Transform a Discrete Value Series}
\usage{
dsct_linear(x, m, b)
}
\arguments{
\item{x}{Discrete value series
(\code{numeric} vector or object of class \code{"discretes"}).}

\item{m}{A numeric value indicating the multiplier.}

\item{b}{A numeric value indicating the intercept. Allowed to be missing
(as opposed to setting equal to zero), to preserve signed zeroes.}
}
\value{
A linearly transformed discretes object.
}
\description{
Apply a linear function to a discrete value series (\code{m * x + b}).
Internal function, with preference to use \code{+}, \code{-}, \code{*}, and \code{/} instead.
}
\examples{
discretes:::dsct_linear(integers(), m = 2)
discretes:::dsct_linear(integers(), m = 2, b = 0.5)
}
