% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsct_negate.R
\name{dsct_negate}
\alias{dsct_negate}
\title{Negate a Discrete Value Series}
\usage{
dsct_negate(x)
}
\arguments{
\item{x}{Discrete value series
(\code{numeric} vector or object of class \code{"discretes"}).}
}
\description{
Negate a Discrete Value Series
}
\seealso{
Other transformations: 
\code{\link{dsct_drop}()},
\code{\link{dsct_invert}()},
\code{\link{dsct_transform}()},
\code{\link{dsct_union}()}
}
\concept{transformations}
