% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsct_power.R
\name{dsct_power}
\alias{dsct_power}
\title{Raise a discrete set to a power.}
\usage{
dsct_power(x, power)
}
\arguments{
\item{x}{Discrete value series
(\code{numeric} vector or object of class \code{"discretes"}).}

\item{power}{The power to raise the series to; numeric of length 0 or 1.
If \code{x} contains negative values, then only odd values are allowed;
an error is thrown otherwise.}
}
\value{
A discretes object where the members are the result of applying the
power transformation \code{^} to \code{x} with the specified \code{power}.
}
\description{
Apply a power transformation to a discrete set \code{x} for a given \code{power};
that is, \code{x^power}.
Internal; use \code{^} operator instead.
}
\examples{
## These are the same
discretes:::dsct_power(natural0(), power = 2)
natural0()^2

## These are also the same
discretes:::dsct_power(integers(), power = 3)
integers()^3
}
