% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsct_union.R
\name{dsct_union}
\alias{dsct_union}
\title{Combine numeric series}
\usage{
dsct_union(...)
}
\arguments{
\item{...}{Objects to combine. Each must be either a numeric series or a
numeric vector.}
}
\value{
A numeric series (inheriting from \code{dsct_union}).
}
\description{
Combine one or more numeric series (or numeric vectors interpreted as
numeric series) into a single union, where each unique discrete value is
counted once.
}
\details{
Values are flattened
}
\note{
While both \code{-0} and \code{+0} can both exist,
}
\examples{
x1 <- natural1()
x2 <- -natural1()
dsct_union(x1, x2)
}
