% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/empty_set.R
\name{empty_series}
\alias{empty_series}
\title{Create an empty numeric series}
\usage{
empty_series(mode = c("double", "integer"))
}
\arguments{
\item{mode}{Character vector of length 1 indicating the numeric type of the
numeric series; either "double" (the default) or "integer".}
}
\value{
An empty numeric series of the specified mode.
}
\description{
Create a numeric series with no discrete values.
}
\examples{
empty_series()
}
