% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/has_discretes.R
\name{has_discretes}
\alias{has_discretes}
\title{Test if values belong to a discrete set}
\usage{
has_discretes(x, values, ..., tol = sqrt(.Machine$double.eps))
}
\arguments{
\item{x}{An discretes object.}

\item{values}{A vector of values to check.}

\item{...}{Reserved for future extensions; must be empty.}

\item{tol}{Numerical tolerance used for snapping \code{values}
to discrete values in the series; single non-negative numeric.}
}
\value{
A logical vector indicating whether each value belongs to the
discrete set defined by \code{x}. \code{NA} values are preserved such that \code{NA} in
\code{values} results in \code{NA} in the output.
}
\description{
Test if values belong to a discrete set
}
\note{
This function does not distinguish between \code{+0} and \code{-0}. For that,
use \code{has_negative_zero()} or \code{has_positive_zero()}.
}
\examples{
has_discretes(natural0(), c(-1, 0, 1, 12.5, NA))
has_discretes(1 / natural1(), 0)
}
