% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integers.R
\name{integers}
\alias{integers}
\title{Integers}
\usage{
integers(from = -Inf, to = Inf)
}
\arguments{
\item{from, to}{Numeric values defining the inclusive range of integers.
Defaults to \code{-Inf} and \code{Inf}, representing all integers.}
}
\value{
An object of class \code{integers_infvctr} (and \code{arithmetic_infvctr},
\code{infvctr}), representing the set of integers within the specified range.
}
\description{
Create an object representing the set of integers within a specified range.
}
\examples{
integers()          # All integers
integers(from = 0) # Non-negative integers
integers(to = 0)   # Non-positive integers
integers(from = -5, to = 5) # Integers from -5
}
