% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_between.R
\name{is_between}
\alias{is_between}
\title{Check if values are between two bounds}
\usage{
is_between(x, lower, upper, include_lower = TRUE, include_upper = TRUE)
}
\arguments{
\item{x}{A numeric vector to be checked.}

\item{lower, upper}{Single numeric values specifying the lower and upper
bounds.}

\item{include_lower, include_upper}{Single logical values indicating whether
the lower and upper bounds are inclusive (closed, \code{TRUE}, the default) or
exclusive (open, \code{FALSE}).}
}
\value{
A logical vector of the same length as \code{x}, with \code{TRUE} for elements
that lie within the specified bounds and \code{FALSE} otherwise. \code{NA} inputs
are propagated to the output.
}
\description{
This function checks if each element of a numeric vector \code{x} lies within the
specified lower and upper bounds. You can choose whether each end is open
or closed.
}
\details{
The convention is adopted that, if \code{lower} and \code{upper} are equal,
both ends of the interval must be closed in order for any value to be
considered between them; otherwise, \code{FALSE} will be returned.
}
\examples{
discretes:::is_between(1:5, lower = 2, upper = 2)
discretes:::is_between(1:5, lower = 2, upper = 2, include_lower = FALSE)
discretes:::is_between(1:5, lower = 2, upper = 4, include_upper = FALSE)
}
