% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_discrete_set.R
\name{is_discrete_set}
\alias{is_discrete_set}
\title{Check if an object is treated as a numeric series}
\usage{
is_discrete_set(x)
}
\arguments{
\item{x}{Object to check.}
}
\value{
\code{TRUE} if \code{x} is treated as a numeric series, \code{FALSE} otherwise.
}
\description{
Returns \code{TRUE} if an object inherits the class \code{"discretes"} or is a
numeric vector.
}
\examples{
is_discrete_set(natural0())
is_discrete_set(c(1, 2, 3))
is_discrete_set("not a numeric series")
}
