% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_discretes.R
\name{is_discretes}
\alias{is_discretes}
\title{Check is if an object is understood as a discretes object}
\usage{
is_discretes(x)
}
\arguments{
\item{x}{Object to check.}
}
\value{
\code{TRUE} if \code{x} can be understood as a "discretes" object,
\code{FALSE} otherwise.
}
\description{
This can happen if an object inherits the class "discretes", or if it
is a numeric vector.
}
\examples{
is_discretes(natural0())
is_discretes(c(1, 2, 3))
is_discretes("not a discretes")
}
