% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iv_invert.R, R/iv_multiply.R, R/iv_negate.R,
%   R/iv_shift.R
\name{iv_invert}
\alias{iv_invert}
\alias{iv_multiply}
\alias{iv_negate}
\alias{iv_shift}
\title{Linear Transformation of Infvctrs}
\usage{
iv_invert(x)

iv_multiply(x, by)

iv_negate(x)

iv_shift(x, by)
}
\arguments{
\item{x}{An infvctr object.}

\item{by}{A numeric value indicating the amount to shift (add) by.}
}
\value{
A linearly transformed infvctr object.
}
\description{
Linear Transformation of Infvctrs
}
\examples{
shift(integers(), 5)
}
