% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math.R
\name{Math.discretes}
\alias{Math.discretes}
\title{Math group generic for numeric series}
\usage{
\method{Math}{discretes}(x, ...)
}
\arguments{
\item{x}{Numeric series
(\code{numeric} vector or object of class \code{"discretes"}).}

\item{...}{Passed to \code{log()} for the \code{base} argument when \code{op == "log"}.}
}
\value{
A numeric series with the math function applied.
}
\description{
Support for \code{exp}, \code{log}, \code{log10}, \code{log2}, and \code{sqrt} on numeric series.
The new series behaves as though the function is applied to each discrete
value in the series.
}
\examples{
exp(integers(from = 0, to = 3))
log(natural1())
sqrt(integers(from = 0, to = 10))
}
