% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/naturalx.R
\name{natural1}
\alias{natural1}
\alias{natural0}
\title{Natural Numbers}
\usage{
natural1()

natural0()
}
\value{
An object of class \code{infvctr}, which is like an infinite-length
vector.
}
\description{
Creates an infinite vector of natural numbers starting at 0 or 1.
}
