% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_discretes.R
\name{new_discretes}
\alias{new_discretes}
\title{Construct an infinite vector object}
\usage{
new_discretes(data, ..., subclass = character())
}
\arguments{
\item{data}{List of fields describing the infinite vector.}

\item{subclass}{Character vector of subclass tags to prepend.}

\item{base}{Optional underlying discretes object being transformed.}
}
\description{
Helper for creating discretes objects that optionally wrap a base
vector.
}
\keyword{internal}
