% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_infvctr.R
\name{new_infvctr}
\alias{new_infvctr}
\title{Construct an infinite vector object}
\usage{
new_infvctr(data, ..., subclass = character())
}
\arguments{
\item{data}{List of fields describing the infinite vector.}

\item{subclass}{Character vector of subclass tags to prepend.}

\item{base}{Optional underlying infvctr being transformed.}
}
\description{
Helper for creating infvctr objects that optionally wrap a base vector.
}
\keyword{internal}
