% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/next_discrete.R, R/prev_discrete.R
\name{next_discrete}
\alias{next_discrete}
\alias{prev_discrete}
\title{Probing discrete values in a distribution}
\usage{
next_discrete(x, from, ..., n = 1L, include_from = TRUE, tol)

prev_discrete(x, from, ..., n = 1L, include_from = TRUE)
}
\arguments{
\item{x}{Vector or discretes.}

\item{from}{Reference value to start searching from; single numeric.}

\item{...}{Reserved for future extensions; must be empty.}

\item{n}{Number of discrete values to find; single positive integer.}

\item{include_from}{Should the \code{from} value be included
in the query? Single logical; defaults to \code{TRUE}.}

\item{tol}{Numerical tolerance used for snapping the \code{from} value
to a discrete value in the series; single non-negative numeric.}
}
\value{
A vector of sequential points starting from \code{from}, which is
included in the vector if it has membership in the series and
\code{include_from = TRUE}. The length of the vector is at most \code{n}, and
will return an empty numeric vector if there is no such discrete
value. \code{next_discrete()} is increasing, while \code{prev_discrete()} is
decreasing, so that earlier values are encountered sooner.
}
\description{
\code{next_discrete()} and \code{prev_discrete()} find the \code{n} discrete values
in a distribution next to a reference point. \code{num_discretes()} finds
the number of discrete values within a range. \code{has_infinite_discretes()}
checks whether there are an infinite amount of discrete values between
a range of values.
}
\note{
Just because there may not be any
}
\examples{
x <- integers(from = 2)
next_discrete(x, from = 1.3)
prev_discrete(x, from = 4, n = 10)
next_discrete(x - 0.7, from = 1.3, n = 4)
}
