% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/num_discretes.R
\name{num_discretes}
\alias{num_discretes}
\title{Number of discrete values in a series}
\usage{
num_discretes(
  x,
  from = -Inf,
  to = Inf,
  ...,
  include_from = TRUE,
  include_to = TRUE,
  tol = sqrt(.Machine$double.eps)
)
}
\arguments{
\item{x}{Numeric series
(\code{numeric} vector or object of class \code{"discretes"}).}

\item{from, to}{Reference values, possibly infinite; both must be length-1
numeric vectors. If \code{to < from}, the range is empty: \code{num_discretes()}
returns \code{0} with a warning.}

\item{...}{Reserved for future extensions; must be empty.}

\item{include_from, include_to}{Should the \code{from} value be included
in the query? Should the \code{to} value? Both must be length-1 logical vectors.}

\item{tol}{Numerical tolerance when checking if a value is in the series.
Single non-negative numeric. See \code{next_discrete()} for details.}
}
\value{
For \code{num_discretes()},
a single non-negative integer, or possibly \code{Inf} for infinitely many
discrete values.
}
\description{
Return the number of discrete values in \code{x} that lie between
\code{from} and \code{to}, or test whether the number of discrete values is infinite.
}
\examples{
num_discretes(-3:3)
num_discretes(c(0.4, 0.4, 0.4, 0))

x <- arithmetic(-3.2, spacing = 0.5)
num_discretes(x)
num_discretes(x, from = -2, to = 2)
num_discretes(1 / x, from = -2, to = 2)
}
