% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/next_discrete.numeric.R
\name{next_discrete.numeric}
\alias{next_discrete.numeric}
\alias{prev_discrete.numeric}
\title{Discrete Values from Numeric Vectors}
\usage{
\method{next_discrete}{numeric}(x, from, ..., n = 1, include_from = FALSE)

\method{prev_discrete}{numeric}(x, from, ..., n = 1, include_from = FALSE)
}
\arguments{
\item{x}{A numeric vector.}

\item{...}{Arguments to pass to other methods.}

\item{n}{Number of discrete values to find.}
}
\value{
A numeric vector.
}
\description{
Discrete Values from Numeric Vectors
}
\examples{
x <- c(1.2, 1.2, 1.2, 5, -2.5, 3.4, 0, 7.8)
next_discrete(x, from = -100)
next_discrete(x, from = -100, n = 4)
next_discrete(x, from = -100, n = 10)
prev_discrete(x, from = 10)
num_discretes(x, from = 0, to = 5)
has_infinite_discretes(x, from = -Inf, to = Inf)
}
