% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.discretes.R
\name{plot.discretes}
\alias{plot.discretes}
\title{Plot Discrete Values}
\usage{
\method{plot}{discretes}(
  x,
  ...,
  from = -Inf,
  to = Inf,
  closeness = 0.01,
  tol = sqrt(.Machine$double.eps)
)
}
\arguments{
\item{x}{Discrete value series
(\code{numeric} vector or object of class \code{"discretes"}).}

\item{...}{Additional arguments passed to the underlying \code{plot()} function.}

\item{from, to}{Numeric values defining the range to plot; single numerics.}

\item{closeness}{Numeric value indicating how close to the (non-infinite)
sinks points should no longer be plotted. This is because there are an
infinite number of points around each sink.}

\item{tol}{Numerical tolerance used for snapping \code{from} or \code{to} values,
including in internal calls; single non-negative numeric.}
}
\value{
Invisibly returns the input \code{x} object after printing a plot
in Base R.
}
\description{
Plot the discrete values represented by a \code{discretes} object within a
specified interval.
}
\details{
Sinks at finite values are indicated by vertical dotted gray lines.
A red tick mark is used to indicate that a finite sink value is part of
the series.

When the series extends to infinity in either direction, three arrows
(\code{<} or \code{>}) are drawn to indicate this. When infinity is part of the series,
the last arrow is red.

This is a simple plotting scheme with naive handling of infinite discrete
values:
\itemize{
\item The \code{closeness} parameter does not adjust with the scale of the data, so
may require tuning more often by the user.
\item When the series extends to infinity (in either direction), an arbitrary
cutoff of 10 units beyond the last finite sink or \code{representative()}
value (whichever is closer to the infinite sink) is used. This can be
manually adjusted by changing the \code{from} and \code{to} parameters.
}
}
\examples{
plot(integers())
plot(integers(), from = -50, to = 50)
plot(0.5^natural1(), closeness = 1e-5)
plot(dsct_union(1 / integers(), 0), closeness = 0.05)
}
