% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pluck_discretes.R
\name{pluck_discretes}
\alias{pluck_discretes}
\title{Extract members of a discrete set}
\usage{
pluck_discretes(x, values, ..., tol = sqrt(.Machine$double.eps))
}
\description{
Extract a finite subset of values from a discrete set by asking for
specific values (\code{pluck_discretes()}) or by setting a range
(\code{pluck_between()}). For \code{pluck_discretes()}, specified values "snap" to
those in the discrete set if they are within \code{tol} of a value.
}
