% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.discretes.R
\name{print.discretes}
\alias{print.discretes}
\title{Print a numeric series}
\usage{
\method{print}{discretes}(x, len = 6, ...)
}
\arguments{
\item{x}{Numeric series
(\code{numeric} vector or object of class \code{"discretes"}).}

\item{len}{Number of discrete values to display.}

\item{...}{Further arguments to pass to downstream methods; currently not
used.}
}
\value{
Invisibly returns the input object \code{x}.
}
\description{
Print a numeric series to the console.
}
\examples{
print(integers())
print(1 / natural1())
print(-1 / natural1())
print(1 / integers())
print(1 / integers(), len = 1)
print(1 / integers(), len = 0)
}
