% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/representative.R
\name{representative}
\alias{representative}
\title{Get a representative discrete value in a numeric series}
\usage{
representative(x)
}
\arguments{
\item{x}{Numeric series
(\code{numeric} vector or object of class \code{"discretes"}).}
}
\value{
A single numeric: a representative discrete value from the
numeric series, with the proper mode.
}
\description{
Get a representative discrete value in a numeric series
}
\examples{
representative(integers())
representative(natural1() + 7)
}
