% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discretes-shift.R
\name{shift}
\alias{shift}
\title{Shift an infvctr}
\usage{
shift(x, by)
}
\arguments{
\item{x}{An infvctr object.}

\item{by}{A numeric value indicating the amount to shift (add) by.}
}
\description{
Shift an infvctr
}
\examples{
shift(integers(), 5)

}
