% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset_discretes.R
\name{[.discretes}
\alias{[.discretes}
\alias{[<-.discretes}
\title{Subset a numeric series by position}
\usage{
\method{[}{discretes}(x, i)

\method{[}{discretes}(x, i) <- value
}
\arguments{
\item{x}{A numeric series (object of class \code{"discretes"}).}

\item{i}{Numeric vector of indices. Omit for the
full series (finite only).}

\item{value}{Replacement value; ignored, because replacement via \verb{[<-} is not
supported.}
}
\value{
A vector of discrete values. When the series has no first element or
too few values for positive \code{i}, R returns NA as for ordinary vectors.
For negative \code{i} or missing \code{i}, the full series is obtained first; infinite
series behaviour defaults to that of \code{get_discretes_in()}.
}
\description{
When a series has a well-defined "first" element (e.g. \code{natural1()} starts at
1), subsetting with \verb{[]} materializes a specified part of the series, and
mirrors the behaviour of base R.
Positive \code{i} returns the discrete values at those positions, and
negative \code{i} tries to return the full series with the specified positions
dropped.
}
\details{
Subsetting via \verb{[]} tries to delegate to base R as quickly as possible by first
materializing the series as a vector, and then conducting the subsetting.
\itemize{
\item If \code{i} is missing or has negative values, subsetting is delegated to the
full series materialized via \code{get_discretes_in()} (if possible).
\item If \code{i} is \code{NULL} or length-0, subsetting is delegated to a representative
value of the series.
\item If \code{i} doesn't have negative values, subsetting is delegated to the series
materialized as far out as needed to cover all \code{i}s, via \code{next_discrete()}
from \code{-Inf}.
}
}
\note{
Unlike base R, the following actions are not supported:
\itemize{
\item Replacement via \verb{[<-} (throws an error).
\item Subsetting by a character vector, as though subsetting by entry names.
\item Subsetting by a logical vector.
}
}
\examples{
natural1()[2]
natural1()[c(1, 3, 5)]
integers(1, 5)[-1]   # full series with first value dropped

# Subsetting from the other side of a sink
x <- 1 / natural1()
x[1:3]   # No such thing as a "first" value; returns NA.
y <- dsct_union(x, -1)
y[1:3]   # "-1" is the 1st value, but no such thing as 2nd or 3rd value.
}
