% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsct_drop.R, R/dsct_keep.R
\name{dsct_drop}
\alias{dsct_drop}
\alias{dsct_keep}
\title{Subsetting Discrete Support Values}
\usage{
dsct_drop(
  x,
  ...,
  from = -Inf,
  to = Inf,
  include_from = TRUE,
  include_to = TRUE
)

dsct_keep(
  x,
  ...,
  from = -Inf,
  to = Inf,
  include_from = TRUE,
  include_to = TRUE
)
}
\arguments{
\item{...}{Reserved for future extensions; must be empty.}

\item{from, to}{Numeric values defining the range to keep; single numerics
with \code{from <= to}.}

\item{include_from, include_to}{Logical values indicating whether the
\code{from} and \code{to} values should be included in the kept range; single
logicals.}
}
\value{
A \code{discretes} object representing the subset of discrete values
within the specified range.
}
\description{
Subsetting Discrete Support Values
}
\examples{
x <- integers(from = -3)
dsct_keep(x, from = -1.5, to = 2.5)
dsct_keep(x, to = 2)
}
