% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_discrete.numeric.R
\name{test_discrete.numeric}
\alias{test_discrete.numeric}
\title{Test membership of values against a vector}
\usage{
\method{test_discrete}{numeric}(x, values, ..., tol = sqrt(.Machine$double.eps))
}
\description{
This function is like \code{values \%in\% x}, testing whether each element of
\code{values} can be found in \code{x}, except is more strict about handling
\code{NA} values, which are interpreted as missing data. See details.
}
\details{
\code{NA} values are treated as missing data, so that if \code{x} contains \code{NA},
it is impossible to know whether any value of \code{values} belongs to \code{x},
and therefore the result is \code{NA} for all elements of \code{values}.
Similarly, if any element of \code{values} is \code{NA}, the result for that
element is \code{NA}, except when \code{x} is the empty set (\code{numeric(0L)}),
in which case we can confidently say that, whatever the missing value is,
it cannot belong to \code{x}, and so the result is \code{FALSE}.

This is unlike Base R, which treats \code{NA} as a distinct element of a set.
As a result, \code{NA \%in\% NA} evaluates to \code{TRUE}, whereas
\code{test_discretes(NA, NA)} evaluates to \code{NA}.
}
