% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsct_transform.R
\name{dsct_transform}
\alias{dsct_transform}
\alias{dsct_transform.discretes}
\title{Monotonically Transform a Discrete Value Series}
\usage{
dsct_transform(x, fun, ...)

\method{dsct_transform}{discretes}(
  x,
  fun,
  inv,
  ...,
  domain = c(-Inf, Inf),
  range = c(-Inf, Inf),
  dir = c("increasing", "decreasing")
)
}
\arguments{
\item{x}{Discrete value series
(\code{numeric} vector or object of class \code{"discretes"}).}

\item{fun, inv}{A vectorized, strictly monotonic function to apply to the
discrete support values, and its inverse, \code{inv}.}

\item{...}{Reserved for future extensions.}

\item{domain, range}{Numeric vectors of length 2, indicating the domain and
range of \code{fun} (that is, the interval on which \code{fun} is valid, and the
interval in which \code{fun} maps to).}

\item{dir}{A string, either "increasing" or "decreasing", indicating the
monotonicity of the function \code{fun}.}
}
\value{
A transformed discretes object.
}
\description{
Apply a function that's either strictly increasing or strictly decreasing
to a numeric series.
}
\details{
Strictly increasing means that for any \code{x1 < x2}, it holds that
\code{fun(x1) < fun(x2)}, for all values on the real line. The function \code{-1/x},
for example, is not strictly increasing: its derivative is increasing,
but switches to smaller values after \code{x = 0}, therefore is not strictly
increasing. Strictly decreasing is the opposite, in that we have
\code{fun(x1) > fun(x2)}.

If a decreasing function is provided, the transformation is negated
internally first, and then transformed with fun(-x).
}
\examples{
dsct_transform(integers(), fun = pnorm, inv = qnorm, range = c(0, 1))
dsct_transform(
  as_discretes(0:3),
  fun = cos,
  inv = acos,
  domain = c(0, pi),
  range = c(-1, 1),
  dir = "decreasing"
)

# For numeric inputs, function is applied directly.
# Other arguments beyond `fun` get absorbed in `...` and are not used.
dsct_transform(0:5, exp)
dsct_transform(0:5, exp, log)
}
