% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zeroes_vector.R
\name{zeroes_vector}
\alias{zeroes_vector}
\title{Construct vector of zeroes}
\usage{
zeroes_vector(x)
}
\arguments{
\item{x}{Numeric series
(\code{numeric} vector or object of class \code{"discretes"}).}
}
\value{
A numeric (double) vector of zeroes, either of length 0, 1, or 2.
}
\description{
Make a numeric (double) vector of zeroes of at most length 2:
contains \code{+0} if the series contains \code{+0}, and
contains \code{-0} if the series contains \code{-0}. Internal; useful for delegating
the behaviour of signed zero to how R does it for numeric vectors.
}
\examples{
(none <- discretes:::zeroes_vector(natural1()))
1 / none

(pos <- discretes:::zeroes_vector(integers()))
1 / pos

(neg <- discretes:::zeroes_vector(arithmetic(-0, 1)))
1 / neg

(posneg <- discretes:::zeroes_vector(dsct_union(integers(), -0)))
1 / posneg
}
