% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.double.R
\name{as.double.discretes}
\alias{as.double.discretes}
\title{Convert a discrete set to a numeric vector}
\usage{
\method{as.double}{discretes}(x, ...)
}
\arguments{
\item{x}{Discrete value series
(\code{numeric} vector or object of class \code{"discretes"}).}

\item{...}{Arguments to pass downstream to \code{as.numeric()} that's called
on the resulting vector of discrete values.}
}
\value{
A numeric vector containing all discrete values in \code{x},
ordered from smallest to largest. Returns \code{numeric(0)} when the interval
contains no discrete values. Numeric outputs are wrapped in \code{as.numeric()}.
}
\description{
Return all discrete values in a "discretes" object, if finite.
Throws an error if infinite.
}
\examples{
as.numeric(integers(-3.5, 10))
}
\seealso{
\code{\link[=get_discretes_in]{get_discretes_in()}}
}
