% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.double.R
\name{as.double.discretes}
\alias{as.double.discretes}
\title{Convert a discrete set to a numeric vector}
\usage{
\method{as.double}{discretes}(x, ...)
}
\arguments{
\item{x}{Invctr object.}

\item{...}{Reserved for future extensions; must be empty.}
}
\value{
A numeric vector containing all discrete values in \code{x},
ordered from smallest to largest. Returns \code{numeric(0)} when the interval
contains no discrete values.
}
\description{
Return all discrete values in a "discretes" object, if finite.
If infinite, throws an error.
}
\examples{
as.numeric(integers(-3.5, 10))
}
