% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_discretes.R
\name{as_discretes}
\alias{as_discretes}
\title{Wrap a numeric vector as a discrete set.}
\usage{
as_discretes(x)
}
\arguments{
\item{x}{A numeric vector with no missing values. May contain \code{Inf} and
\code{-Inf}.}
}
\value{
A "discretes" object whose members are the unique values of \code{x}.
}
\description{
The discretes package is careful to distinguish between how a discrete set
is encoded vs. how it's expressed (through, for example, \code{next_discrete()}).
This function allows a discrete set to be encoded by a numeric vector.
}
\details{
This function is useful, for example, when doing
\code{0 * dsct_union(integers(), -0)} -- it would be awkward to simplify this
to \code{c(0, -0)}, yet both signed zeroes would need to be encoded in the
resulting discrete set.

This function also ensures that manipulations remain pure: manipulating
a discrete set returns a new discrete set.
}
\examples{
as_discretes(0:10)

}
