% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsct_linear.R
\name{dsct_linear}
\alias{dsct_linear}
\title{Linearly transform a numeric series}
\usage{
dsct_linear(x, m, b = NULL)
}
\arguments{
\item{x}{Numeric series
(\code{numeric} vector or object of class \code{"discretes"}).}

\item{m}{A numeric value indicating the multiplier.}

\item{b}{A numeric value indicating the intercept. Allowed to be \code{NULL}
(the default) to indicate the transformation \code{m * x} without an intercept,
which is important to distinguish signed zero.}
}
\value{
A numeric series with the linear transformation applied.
}
\description{
Apply a linear function to a numeric series (\code{m * x + b}).
Internal function; prefer using \code{+}, \code{-}, \code{*}, and \code{/} instead.
}
\examples{
discretes:::dsct_linear(integers(), m = 2)
discretes:::dsct_linear(integers(), m = 2, b = 0.5)
}
