% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsct_transform.R
\name{dsct_transform}
\alias{dsct_transform}
\title{Monotonically Transform a Discrete Value Series}
\usage{
dsct_transform(x, fun, inv, domain = c(-Inf, Inf), range = c(-Inf, Inf))
}
\arguments{
\item{x}{Discrete value series
(\code{numeric} vector or object of class \code{"discretes"}).}

\item{fun, inv}{A vectorized, strictly increasing function to apply to the
discrete support values, and its inverse, \code{inv}.}

\item{domain, range}{Numeric vectors of length 2, indicating the domain and
range of \code{fun} (that is, the interval on which \code{fun} is valid, and the
interval in which \code{fun} maps to).}
}
\value{
A transformed discretes object.
}
\description{
Monotonically Transform a Discrete Value Series
}
\note{
The onus is on the user to ensure that \code{inv} is indeed the inverse of
\code{fun}, that both are vectorized, and that both are strictly increasing.

By strictly increasing, we mean that for any \code{x1 < x2}, it holds that
\code{fun(x1) < fun(x2)}, for all values on the real line. The function \code{-1/x},
for example, is not strictly increasing: its derivative is increasing,
but switches to smaller values after \code{x = 0}, therefore is not strictly
increasing. This becomes important when distinguishing between \code{+0}
and \code{-0}, which has special treatment with \code{dsct_inverse()}, but not
for \code{dsct_transform()}.
}
\examples{
dsct_transform(integers(), fun = exp, inv = log)
}
\seealso{
Other transformations: 
\code{\link{dsct_drop}()},
\code{\link{dsct_invert}()},
\code{\link{dsct_negate}()},
\code{\link{dsct_union}()}
}
\concept{transformations}
