% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/empty_set.R
\name{empty_series}
\alias{empty_series}
\title{Create an empty discrete set}
\usage{
empty_series(mode = c("double", "integer"))
}
\arguments{
\item{mode}{Character vector of length 1 indicating the numeric type of the
discrete set; either "double" (the default) or "integer".}
}
\value{
An empty discrete set of the specified mode.
}
\description{
Create a discrete set with no members.
}
\examples{
empty_series()
}
