% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsct_exp.R
\name{dsct_raise}
\alias{dsct_raise}
\alias{dsct_exp}
\title{Exponentiation of a discrete set}
\usage{
dsct_raise(x, base = exp(1))

dsct_exp(x)
}
\arguments{
\item{x}{Discrete value series
(\code{numeric} vector or object of class \code{"discretes"}).}

\item{base}{The base to use for exponentiation; numeric of length 0 or 1,
with negative values not allowed (an error is thrown otherwise).}
}
\value{
A discretes object where each member is the result of
exponentiating a member of \code{x} by \code{base} (that is, \code{base^x}).
}
\description{
Exponentiate a discrete set \code{x} with a given base; that is, \code{base^x}.
\code{dsct_exp()} is a special case of \code{dsct_raise()} where the base is \code{exp(1)}.
Internal; use \code{exp()} or \code{^} operators instead.
}
\examples{
# These are the same
discretes:::dsct_exp(integers())
discretes:::dsct_raise(integers(), base = exp(1))
exp(integers())

# These are also the same
discretes:::dsct_raise(integers(), base = 2)
2^integers()

# This also works. Notice how the set reduces.
0^integers()
}
