% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_discretes_at.R, R/get_discretes_in.R
\name{get_discretes_at}
\alias{get_discretes_at}
\alias{get_discretes_in}
\title{Extract members of a discrete set}
\usage{
get_discretes_at(x, values, ..., tol = sqrt(.Machine$double.eps))

get_discretes_in(
  x,
  ...,
  from = -Inf,
  to = Inf,
  include_from = TRUE,
  include_to = TRUE,
  tol = sqrt(.Machine$double.eps)
)
}
\arguments{
\item{x}{Invctr object.}

\item{values}{Numeric vector of values to pull from the discrete set \code{x}.}

\item{...}{Reserved for future extensions; must be empty.}

\item{tol}{Numerical tolerance used for snapping \code{from} or \code{to} values
to a discrete value in the series; single non-negative numeric.}

\item{from, to}{Reference values, possibly infinite. \code{from} must be less than
or equal to \code{to}; both must be length-1 numeric vectors.}

\item{include_from, include_to}{Should the \code{from} value be included
in the query? Should the \code{to} value? Both must be length-1 logical vectors.}
}
\value{
A numeric vector containing all discrete values between \code{from} and
\code{to}. Returns \code{numeric(0)} when the interval contains no discrete values.
}
\description{
Extract a finite subset of values from a discrete set by asking for
specific values (\code{get_discretes_at()}) or by setting a range
(\code{pluck_between()}). For \code{get_discretes_at()}, specified values "snap" to
those in the discrete set if they are within \code{tol} of a value.
}
\examples{
get_discretes_in(integers(), from = 6.6, to = 10.1)
get_discretes_in(1 / integers(1, 4, n_left = 3, n_right = 5))
}
