% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_discretes_at.R, R/get_discretes_in.R
\name{get_discretes_at}
\alias{get_discretes_at}
\alias{get_discretes_in}
\title{Extract discrete values from a numeric series}
\usage{
get_discretes_at(x, values, ..., tol = sqrt(.Machine$double.eps))

get_discretes_in(
  x,
  from = -Inf,
  to = Inf,
  ...,
  include_from = TRUE,
  include_to = TRUE,
  tol = sqrt(.Machine$double.eps)
)
}
\arguments{
\item{x}{Numeric series
(\code{numeric} vector or object of class \code{"discretes"}).}

\item{values}{Numeric vector of values to pull from the numeric series \code{x}.}

\item{...}{Reserved for future extensions; must be empty.}

\item{tol}{Numerical tolerance when testing membership. Single non-negative
numeric. See \code{\link[=next_discrete]{next_discrete()}} for details.}

\item{from, to}{Reference values, possibly infinite. \code{from} must be less than
or equal to \code{to}; both must be length-1 numeric vectors.}

\item{include_from, include_to}{Should the \code{from} value be included
in the query? Should the \code{to} value? Both must be length-1 logical vectors.}
}
\value{
A numeric vector containing all discrete values in the provided series \code{x}:
\itemize{
\item For \code{get_discretes_in()}, all discrete values between \code{from} and \code{to},
ordered from smallest to largest.
\item For \code{get_discretes_at()}, the discrete values that the given \code{values} are
members of. Membership is tested with \code{tol} at the underlying series,
returning the discrete values in the series rather than the supplied
\code{values}. Discrete values not within \code{tol} of the supplied \code{values}
are dropped.
}

An error will be thrown in \code{get_discretes_in()} if there are infinitely
many points in the range.
}
\description{
Extract a finite subset of discrete values from a numeric series by asking
for specific values (\code{get_discretes_at()}) or by setting a range
(\code{get_discretes_in()}). For \code{get_discretes_at()}, each value in \code{values} is
tested for membership using \code{tol} (passed down to the underlying series); if
it is a member, the
corresponding discrete value is returned, otherwise it is dropped. \code{NA}
values are kept in place.
}
\examples{
get_discretes_in(integers(), from = 6.6, to = 10.1)
get_discretes_in(1 / arithmetic(1, 4, n_left = 3, n_right = 5))
get_discretes_at(integers(), values = c(-10, 4, 3.5, 10, NA))
get_discretes_at(integers(), values = 5.5)
}
\seealso{
\code{\link[=as.double.discretes]{as.double.discretes()}}
}
