% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/has_discretes.R
\name{has_discretes}
\alias{has_discretes}
\title{Check which values are in a numeric series}
\usage{
has_discretes(x, values, ..., tol = sqrt(.Machine$double.eps))
}
\arguments{
\item{x}{Numeric series
(\code{numeric} vector or object of class \code{"discretes"}).}

\item{values}{A vector of values to check.}

\item{...}{Reserved for future extensions; must be empty.}

\item{tol}{Numerical tolerance when checking if a value is in the series.
Single non-negative numeric. See \code{next_discrete()} for details.}
}
\value{
A logical vector indicating whether each value is in the numeric
series \code{x}. \code{NA} values are preserved such that \code{NA} in \code{values} results
in \code{NA} in the output.
}
\description{
Check which values are in a numeric series
}
\note{
This function does not distinguish between \code{+0} and \code{-0}. For that,
use \code{has_negative_zero()} or \code{has_positive_zero()}.
}
\examples{
has_discretes(natural0(), c(-1, 0, 1, 12.5, NA))
has_discretes(1 / natural1(), 0)
}
