% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/has_sink.R
\name{has_sink_in}
\alias{has_sink_in}
\alias{has_sink_at}
\title{Test for sinks in a numeric series}
\usage{
has_sink_in(x, from = -Inf, to = Inf)

has_sink_at(x, value, dir = c("either", "left", "right", "both"))
}
\arguments{
\item{x}{Numeric series
(\code{numeric} vector or object of class \code{"discretes"}).}

\item{from, to}{Reference values, possibly infinite. \code{from} must be less than
or equal to \code{to}; both must be length-1 numeric vectors.}

\item{value}{Single numeric to check for the existence of a sink.}

\item{dir}{Single character: direction of the sink. \code{"either"} (default)
ignores direction; \code{"left"} or \code{"right"} require the sink to be
approached from that side; \code{"both"} requires it to be approached from
both sides.}
}
\value{
A length-one logical: \code{TRUE} if a sink exists in the range
(\code{has_sink_in}) or at \code{value} with the given \code{dir}
(\code{has_sink_at}); \code{FALSE} otherwise.
}
\description{
\code{has_sink_in()} tests whether a numeric series has a sink in the interval
\verb{[from, to]}. \code{has_sink_at()} tests whether there is a sink at a
given value, optionally with a specific direction.
}
\examples{
# The set of integers have sinks at +Inf and -Inf
has_sink_in(integers())
has_sink_at(integers(), Inf)
has_sink_at(integers(), -Inf, dir = "right")

# The set 1, 1/2, 1/4, 1/8, ... has a sink at 0 approached from the right.
halves <- 0.5^natural0()
has_sink_in(halves, to = 0)
has_sink_at(halves, 0, dir = "right")

# Reciprocal of integers: sink at 0 from both sides
reciprocals <- 1 / integers()
has_sink_at(reciprocals, 0, dir = "both")
}
\seealso{
See \code{\link[=sinks]{sinks()}} to get all sinks, and a description of sinks.
}
