% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integers.R
\name{integers}
\alias{integers}
\title{Integer Discrete Values}
\usage{
integers(from = -Inf, to = Inf)
}
\arguments{
\item{from, to}{Numeric values defining the inclusive range of integers.
Defaults to \code{-Inf} and \code{Inf}, representing all integers.}
}
\value{
A "discretes" object of class \code{dsct_integers},
representing the set of integers within the specified range.
}
\description{
Create a "discretes" object representing the set of integers within a
specified range.
}
\examples{
integers()                  # All integers
integers(from = 0)          # Non-negative integers
integers(to = 1.5)          # Ends at 1.
integers(from = -5, to = 5) # Integers from -5 to 5.
}
