% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integers.R, R/naturalx.R
\name{integers}
\alias{integers}
\alias{natural1}
\alias{natural0}
\title{Integer Discrete Value Series}
\usage{
integers(from = -Inf, to = Inf)

natural1()

natural0()
}
\arguments{
\item{from, to}{Numeric values defining the range of integers.
Defaults to \code{-Inf} and \code{Inf}, representing all integers, although
is not closed: \code{-Inf} and \code{Inf} are never included in the series.}
}
\value{
An arithmetic discrete value series (inheriting class
\code{"dsct_arithmetic"}) representing the set of integers within the specified
range.
}
\description{
Use \code{integers()} to create a discrete value series consisting of integers
within a specified range, possibly unbounded on either end. Use \code{natural0()}
and \code{natural1()} for the set of natural numbers starting at either 0 or 1.
}
\examples{
integers()                  # All integers
integers(from = 0)          # Non-negative integers
integers(to = 1.5)          # Ends at 1.
integers(-5, 5) # Integers from -5 to 5.
natural1()
natural0()

# Infinity is never contained in the series.
test_discrete(integers(), Inf)
}
\seealso{
\code{\link[=arithmetic]{arithmetic()}}
}
