% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_discrete_set.R
\name{is_discrete_set}
\alias{is_discrete_set}
\title{Check is if an object is understood as a discretes object}
\usage{
is_discrete_set(x)
}
\arguments{
\item{x}{Object to check.}
}
\value{
\code{TRUE} if \code{x} can be understood as a "discretes" object,
\code{FALSE} otherwise.
}
\description{
This can happen if an object inherits the class "discretes", or if it
is a numeric vector.
}
\examples{
is_discrete_set(natural0())
is_discrete_set(c(1, 2, 3))
is_discrete_set("not a discretes")
}
