% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iv_invert.R, R/iv_invert.infvctr.R,
%   R/iv_linear.R, R/iv_linear.linear_transform.R, R/iv_negate.R,
%   R/iv_negate.arithmetic_infvct.R, R/iv_negate.infvctr.R
\name{iv_invert}
\alias{iv_invert}
\alias{iv_invert.infvctr}
\alias{iv_linear}
\alias{iv_linear.linear_transform}
\alias{iv_negate}
\alias{iv_negate.arithmetic_infvctr}
\alias{iv_negate.infvctr}
\title{Linear Transformation of Infvctrs}
\usage{
iv_invert(x)

\method{iv_invert}{infvctr}(x)

iv_linear(x, m, b)

\method{iv_linear}{linear_transform}(x, m, b)

iv_negate(x)

\method{iv_negate}{arithmetic_infvctr}(x)

\method{iv_negate}{infvctr}(x)
}
\arguments{
\item{x}{An infvctr object.}

\item{m}{A numeric value indicating the multiplier.}

\item{b}{A numeric value indicating the bias (addend).}
}
\value{
A linearly transformed infvctr object.
}
\description{
Linear Transformation of Infvctrs
}
\section{Methods (by generic)}{
\itemize{
\item \code{iv_linear(linear_transform)}: Compose linear transforms.

}}
\section{Functions}{
\itemize{
\item \code{iv_invert()}: Take element-wise reciprocals of an infvctr.

\item \code{iv_invert(infvctr)}: Default reciprocal transform for infvctrs.

\item \code{iv_negate(arithmetic_infvctr)}: Negate an arithmetic progression.

\item \code{iv_negate(infvctr)}: Negate an infvctr.

}}
\examples{
iv_linear(integers(), m = 2, b = 3)
}
