% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsct_linear.R, R/dsct_linear.dsct_linear.R,
%   R/dsct_negate.R, R/dsct_negate.dsct_arithmetic.R
\name{dsct_linear}
\alias{dsct_linear}
\alias{dsct_linear.dsct_linear}
\alias{dsct_negate}
\alias{dsct_negate.discretes}
\alias{dsct_negate.dsct_arithmetic}
\title{Linear Transformation of discretes}
\usage{
dsct_linear(x, m, b)

\method{dsct_linear}{dsct_linear}(x, m, b)

dsct_negate(x)

\method{dsct_negate}{discretes}(x)

\method{dsct_negate}{dsct_arithmetic}(x)
}
\arguments{
\item{x}{An discretes object.}

\item{m}{A numeric value indicating the multiplier.}

\item{b}{A numeric value indicating the bias (addend).}
}
\value{
A linearly transformed discretes object.
}
\description{
Linear Transformation of discretes
}
\section{Functions}{
\itemize{
\item \code{dsct_linear(dsct_linear)}: Compose linear transforms.

\item \code{dsct_negate(discretes)}: Negate a discretes object.

\item \code{dsct_negate(dsct_arithmetic)}: Negate an arithmetic progression.

}}
\examples{
dsct_linear(integers(), m = 2, b = 3)
}
