% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.discretes.R
\name{plot.discretes}
\alias{plot.discretes}
\title{Plot a numeric series}
\usage{
\method{plot}{discretes}(
  x,
  from = -Inf,
  to = Inf,
  ...,
  closeness = 0.01,
  tol = sqrt(.Machine$double.eps)
)
}
\arguments{
\item{x}{Numeric series
(\code{numeric} vector or object of class \code{"discretes"}).}

\item{from, to}{Numeric values defining the range to plot; single numerics.}

\item{...}{Additional arguments passed to the underlying \code{plot()} function.}

\item{closeness}{Numeric value indicating how close to the (non-infinite)
sinks the points should no longer be plotted. This is because there are an
infinite number of points around each sink.}

\item{tol}{Passed to \code{dsct_keep()} when subsetting the series between \code{from}
and \code{to}.}
}
\value{
Invisibly returns the input \code{x} object after printing a plot
in Base R.
}
\description{
Plot the discrete values in a numeric series within a specified interval.
}
\details{
Sinks at finite values are indicated by vertical dotted gray lines.
A red tick mark is used to indicate that a finite sink value is part of
the series.

When the series extends to infinity in either direction, three arrows
(\code{<} or \code{>}) are drawn to indicate this. When infinity is part of the series,
the last arrow is red.

This is a simple plotting scheme with naive handling of infinite discrete
values:
\itemize{
\item The \code{closeness} parameter does not adjust with the scale of the data, so
may require tuning more often by the user.
\item When the series extends to infinity (in either direction), an arbitrary
cutoff of 10 units beyond the last finite sink or \code{representative()}
value (whichever is closer to the infinite sink) is used. This can be
manually adjusted by changing the \code{from} and \code{to} parameters.
}
}
\note{
If this function takes a long time to plot, it's likely because
your series has "slowly varying" behaviour next to a sink,
like \code{1 / natural1()}, where discrete values pile up rapidly
while approaching the sink very slowly. To avoid so many points from
being plotted, increase the \code{closeness} argument.
}
\examples{
plot(integers())
plot(integers(), from = -50, to = 50)
plot(0.5^natural1(), closeness = 1e-3)
plot(dsct_union(0.5^natural1(), 0), closeness = 1e-3)
}
