% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/representative.R
\name{representative}
\alias{representative}
\title{Get a representative value in a discrete set}
\usage{
representative(x)
}
\arguments{
\item{x}{A discretes object.}
}
\value{
A single numeric vector with a representative value from the
discrete set, with the proper mode.
}
\description{
Get a representative value in a discrete set
}
\examples{
representative(integers())
representative(natural1() + 7)
}
