% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sinks.R
\name{sinks}
\alias{sinks}
\alias{sink_is_between}
\title{Sinks}
\usage{
sinks(x)

sink_is_between(x, from = -Inf, to = Inf, ..., tol = sqrt(.Machine$double.eps))
}
\arguments{
\item{x}{Discrete value series
(\code{numeric} vector or object of class \code{"discretes"}).}

\item{from}{Reference value to start searching from; single numeric.}

\item{...}{Reserved for future extensions; must be empty.}

\item{tol}{Numerical tolerance used for snapping the \code{from} value
to a discrete value in the series; single non-negative numeric.}
}
\value{
A matrix with two columns: \code{location} and \code{direction}. Each row
corresponds to a sink, where \code{location} is the location of the sink
(possibly infinite), and \code{direction} indicates whether the sink is
approached from the left (-1) or right (1).
}
\description{
Sinks are limit points in a discrete set. That means that discrete values
get arbitrarily close to the sink (from the left or right), and that there
are infinitely many discrete points around the sink.
}
\examples{
# The set of integers have sinks at +Inf and -Inf
sinks(integers())
sink_is_between(integers())

# The set 1, 1/2, 1/4, 1/8, ... has a sink at 0 approached from the right.
halves <- 0.5^natural0()
sinks(halves)
sink_is_between(halves, to = 0)
sink_is_between(halves, from = 0)

# The reciprocal of the integers has a sink at 0 approached from both the
# left and right; while the integer 0 gets mapped to Inf, infinity is not a
# sink because discrete values don't get arbitrarily close to it.
reciprocals <- 1 / integers()
sinks(reciprocals)
test_discrete(reciprocals, Inf) # Yet Inf is a member.
}
