% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sinks.R
\name{sinks}
\alias{sinks}
\alias{has_sink}
\title{Sinks}
\usage{
sinks(x)

has_sink(x, from = -Inf, to = Inf)
}
\arguments{
\item{x}{Discrete value series
(\code{numeric} vector or object of class \code{"discretes"}).}

\item{from, to}{Reference values, possibly infinite. \code{from} must be less than
or equal to \code{to}; both must be length-1 numeric vectors.}
}
\value{
A matrix with two columns: \code{location} and \code{direction}. Each row
corresponds to a sink, where \code{location} is the location of the sink
(possibly infinite), and \code{direction} indicates whether the sink is
approached from the left (-1) or right (1).
}
\description{
Sinks are limit points in a discrete set. That means that discrete values
get arbitrarily close to the sink (from the left or right), and that there
are infinitely many discrete points around the sink.
}
\examples{
# The set of integers have sinks at +Inf and -Inf
sinks(integers())
has_sink(integers())

# The set 1, 1/2, 1/4, 1/8, ... has a sink at 0 approached from the right.
halves <- 0.5^natural0()
sinks(halves)
has_sink(halves, to = 0)
has_sink(halves, from = 0)

# The reciprocal of the integers has a sink at 0 approached from both the
# left and right; while the integer 0 gets mapped to Inf, infinity is not a
# sink because discrete values don't get arbitrarily close to it.
reciprocals <- 1 / integers()
sinks(reciprocals)
has_discretes(reciprocals, Inf) # Yet Inf is a member.
}
