% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_discrete.R,
%   R/test_discrete.arithmetic_infvctr.R, R/test_discrete.inverse.R,
%   R/test_discrete.linear_transform.R, R/test_discrete.negate_infvctr.R
\name{test_discrete}
\alias{test_discrete}
\alias{test_discrete.arithmetic_infvctr}
\alias{test_discrete.inverse}
\alias{test_discrete.linear_transform}
\alias{test_discrete.negate_infvctr}
\title{Test if values belong to a discrete set}
\usage{
test_discrete(x, values)

\method{test_discrete}{arithmetic_infvctr}(x, values, ..., tol = 1e-12)

\method{test_discrete}{inverse}(x, values, ...)

\method{test_discrete}{linear_transform}(x, values, ...)

\method{test_discrete}{negate_infvctr}(x, values, ...)
}
\arguments{
\item{x}{An infvctr object.}

\item{values}{A vector of values to check.}

\item{tol}{Numerical tolerance used when checking whether values align with
the progression.}
}
\value{
A logical vector indicating whether each value belongs to the
discrete set defined by \code{x}. \code{NA} values are preserved such that \code{NA} in
\code{values} results in \code{NA} in the output.
}
\description{
Test if values belong to a discrete set
}
\section{Methods (by class)}{
\itemize{
\item \code{test_discrete(arithmetic_infvctr)}: Membership test for arithmetic progressions.

\item \code{test_discrete(inverse)}: Membership test for reciprocal transforms.

\item \code{test_discrete(linear_transform)}: Membership test for linear transforms.

\item \code{test_discrete(negate_infvctr)}: Membership test for negated infvctr objects.

}}
\examples{
test_discrete(natural0(), c(-1, 0, 1, 12.5, NA))
test_discrete(1 / natural1(), 0)
}
