% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_discrete.R,
%   R/test_discrete.dsct_arithmetic.R, R/test_discrete.dsct_inverse.R,
%   R/test_discrete.dsct_linear.R, R/test_discrete.dsct_negation.R
\name{test_discrete}
\alias{test_discrete}
\alias{test_discrete.dsct_arithmetic}
\alias{test_discrete.dsct_inverse}
\alias{test_discrete.dsct_linear}
\alias{test_discrete.dsct_negate}
\title{Test if values belong to a discrete set}
\usage{
test_discrete(x, values)

\method{test_discrete}{dsct_arithmetic}(x, values, ..., tol = 1e-12)

\method{test_discrete}{dsct_inverse}(x, values, ...)

\method{test_discrete}{dsct_linear}(x, values, ...)

\method{test_discrete}{dsct_negate}(x, values, ...)
}
\arguments{
\item{x}{An discretes object.}

\item{values}{A vector of values to check.}

\item{tol}{Numerical tolerance used when checking whether values align with
the progression.}
}
\value{
A logical vector indicating whether each value belongs to the
discrete set defined by \code{x}. \code{NA} values are preserved such that \code{NA} in
\code{values} results in \code{NA} in the output.
}
\description{
Test if values belong to a discrete set
}
\section{Methods (by class)}{
\itemize{
\item \code{test_discrete(dsct_arithmetic)}: Membership test for arithmetic progressions.

\item \code{test_discrete(dsct_inverse)}: Membership test for reciprocal transforms.

\item \code{test_discrete(dsct_linear)}: Membership test for linear transforms.

\item \code{test_discrete(dsct_negate)}: Membership test for negated discrete objects.

}}
\examples{
test_discrete(natural0(), c(-1, 0, 1, 12.5, NA))
test_discrete(1 / natural1(), 0)
}
