% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_discrete.R
\name{test_discrete}
\alias{test_discrete}
\title{Test if values belong to a discrete set}
\usage{
test_discrete(x, values)
}
\arguments{
\item{x}{An discretes object.}

\item{values}{A vector of values to check.}
}
\value{
A logical vector indicating whether each value belongs to the
discrete set defined by \code{x}. \code{NA} values are preserved such that \code{NA} in
\code{values} results in \code{NA} in the output.
}
\description{
Test if values belong to a discrete set
}
\examples{
test_discrete(natural0(), c(-1, 0, 1, 12.5, NA))
test_discrete(1 / natural1(), 0)
}
